Fixed the tests to make them pass
The steps done to fix the tests are as following:

The first thing was to make the embedding of the Client
structs more clear by importing the katzenpost/client as katzenClient
since it seems that in some occasions it was getting overwritten. The
embedded katzenpost/client.Client was also initialized in the New function
and the exterior Meson-client.Client was intialized following the patterns
of the former.

The second thing done to make the tests pass was to change the expected values.
In the first test the client was not being started with the Start() function. In the second function the expect values got changed from gor to +gor and the reply had the trailling zeroes removed.